rewrite for withPlaceholder
