+++ packages/react/src/components/Placeholder/Placeholder.tsx
@@ -5,11 +5,14 @@ import { withComponentMap } from '../../enhancers/withComponentMap';
 import { PagesEditor } from '@sitecore-content-sdk/content/editing';
 import { withSitecore } from '../../enhancers/withSitecore';
 import {
+  getComponentForRendering,
   getPlaceholderRenderings,
-  getRenderedComponents,
+  getRenderedComponentProps,
   renderEmptyPlaceholder,
 } from './placeholder-utils';
 import ErrorBoundary from '../ErrorBoundary';
+import { ComponentRendering } from '@sitecore-content-sdk/content/layout';
+import { PlaceholderMetadata } from './PlaceholderMetadata';
 
 const PlaceholderComponent = (props: PlaceholderProps) => {
   const renderingData = props.rendering;
@@ -27,9 +30,9 @@ const PlaceholderComponent = (props: PlaceholderProps) => {
     }
   }, [isEmpty]); // Empty array = runs once on mount
 
-  const renderChildren = () => {
-    const childProps: PlaceholderProps = { ...props };
-
+  const renderPlhChildren = () => {
+    const childProps = { ...props };
+    // TODO: cleanup more props
     delete childProps.componentMap;
 
     const components = getRenderedComponents(props, placeholderRenderings);
@@ -56,13 +59,91 @@ const PlaceholderComponent = (props: PlaceholderProps) => {
   };
 
   // Using error boundary for errors that may happen within Placeholder itself
-  return <ErrorBoundary errorComponent={props.errorComponent}>{renderChildren()}</ErrorBoundary>;
+  return <ErrorBoundary errorComponent={props.errorComponent}>{renderPlhChildren()}</ErrorBoundary>;
 };
 
-const PlaceholderWithComponentMap = withComponentMap(PlaceholderComponent);
+/**
+ * Renders the components for the placeholder based on the provided rendering data.
+ * @param {PlaceholderProps} props placeholder component props
+ * @param {ComponentRendering[]} placeholderRenderings renderings within placeholder
+ * @returns {React.ReactNode | React.ReactElement[]} rendered components
+ */
+export const getRenderedComponents = (
+  props: PlaceholderProps,
+  placeholderRenderings: ComponentRendering[]
+) => {
+  const { name, missingComponentComponent, hiddenRenderingComponent } = props;
+
+  const transformedComponents = placeholderRenderings
+    .map((componentRendering: ComponentRendering, index: number) => {
+      const component = getComponentForRendering(
+        componentRendering,
+        name,
+        props.componentMap,
+        hiddenRenderingComponent,
+        missingComponentComponent
+      );
+      const key = componentRendering.uid || `component-${index}`;
+
+      const renderedProps = getRenderedComponentProps(props, componentRendering, key);
+      const finalRenderedProps = props.modifyComponentProps
+        ? props.modifyComponentProps(renderedProps)
+        : renderedProps;
+
+      let rendered = React.createElement<{ [attr: string]: unknown }>(
+        component.component as React.ComponentType,
+        finalRenderedProps
+      );
+
+      if (!component.isEmpty) {
+        const errorBoundaryKey = rendered.type + '-' + index;
+
+        const disableSuspense = props.disableSuspense || false;
+        rendered = (
+          <ErrorBoundary
+            data-testid="error-boundary"
+            key={errorBoundaryKey}
+            errorComponent={props.errorComponent}
+            componentLoadingMessage={props.componentLoadingMessage}
+            isDynamic={component.dynamic}
+            disableSuspense={disableSuspense}
+            rendering={rendered.props.rendering as ComponentRendering}
+          >
+            {rendered}
+          </ErrorBoundary>
+        );
+      }
+
+      // if in edit mode then emit shallow chromes for hydration in Pages
+      if (props.page.mode.isEditing) {
+        return (
+          <PlaceholderMetadata key={key} rendering={componentRendering}>
+            {rendered}
+          </PlaceholderMetadata>
+        );
+      }
+
+      return rendered;
+    })
+    .filter((element) => element); // remove nulls
+
+  if (props.page.mode.isEditing) {
+    return [
+      <PlaceholderMetadata
+        key={(props.rendering as ComponentRendering).uid}
+        placeholderName={name}
+        rendering={props.rendering as ComponentRendering}
+      >
+        {transformedComponents}
+      </PlaceholderMetadata>,
+    ];
+  }
+
+  return transformedComponents;
+};
 
 /**
  * The Placeholder component.
  * @public
  */
-export const Placeholder = withSitecore()(PlaceholderWithComponentMap);
+export const Placeholder = withSitecore()(withComponentMap(PlaceholderComponent));
