+++ packages/react/src/components/Placeholder/index.ts
@@ -1,4 +1,4 @@
-export { Placeholder, PlaceholderComponent } from './Placeholder';
+export { Placeholder, getRenderedComponents } from './Placeholder';
 export { PlaceholderMetadata } from './PlaceholderMetadata';
 export { PlaceholderProps, AppPlaceholderProps } from './models';
 export { AppPlaceholder } from './AppPlaceholder';
